requirements.txt
openai>=1.30.0
langgraph>=0.2.0
python-dotenv>=1.0.0
rich>=13.0.0
pypdf>=4.0.0
